restored the bottom tab
