+++ app/_layout.tsx
@@ -1,7 +1,50 @@
 //Top Level Layout
-import { Stack } from "expo-router";
+// import { Stack } from "expo-router";
+
+// export default function RootLayout() {
+//   return <Stack />;
+// }
+// app/_layout.tsx
+import { Slot, useRouter, useSegments } from 'expo-router';
+import { auth } from '@/lib/firebase-config';
+import { onAuthStateChanged } from 'firebase/auth';
+import React, { useEffect, useState } from 'react';
+import { ActivityIndicator, View } from 'react-native';
 
 export default function RootLayout() {
-  return <Stack />;
+  const [checking, setChecking] = useState(true);
+  const [isLoggedIn, setIsLoggedIn] = useState(false);
+  const segments = useSegments();
+  const router = useRouter();
+
+  useEffect(() => {
+    const unsubscribe = onAuthStateChanged(auth, (user) => {
+      setIsLoggedIn(!!user);
+      setChecking(false);
+    });
+
+    return () => unsubscribe();
+  }, []);
+
+  useEffect(() => {
+    const inAuthGroup = segments[0] !== '(tabs)';
+    if (!checking) {
+      if (!isLoggedIn && !inAuthGroup) {
+        router.replace('/login');
+      } else if (isLoggedIn && inAuthGroup) {
+        router.replace('/(tabs)/feed');
+      }
+    }
+  }, [checking, isLoggedIn, segments]);
+
+  if (checking) {
+    return (
+      <View style={{ flex: 1, justifyContent: 'center', alignItems: 'center' }}>
+        <ActivityIndicator size="large" />
+      </View>
+    );
+  }
+
+  return <Slot />;
 }
